requirements.txt
fastapi
uvicorn[standard]
python-dotenv
tortoise-orm
passlib[bcrypt]
aiofiles
aerich
python-multipart
